% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{align_cats}
\alias{align_cats}
\title{Align Cats}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{img}{the filepath to the image}
  \item{alignment}{the D&D alignment of the cat in question}
}
}
\source{
Photos of cats taken by Afton Coombs
}
\usage{
align_cats
}
\description{
This is a toy dataset you can use to experiment with the `ggalignment`
package. The dataset contains pictures of the author's cats. :)
}
\keyword{datasets}
