% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alignment_vals}
\alias{alignment_vals}
\title{Alignment Values}
\format{
A data.frame vector containing 1 column of 9 elements, each one a possible
alignment
\describe{
\item{alignment}{the nine possible alignments}
}
}
\source{
\url{https://dungeonsdragons.fandom.com/wiki/Alignment}
}
\usage{
alignment_vals
}
\description{
A vector of possible alignment values.
}
\keyword{datasets}
