% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cats.R
\name{example_cats}
\alias{example_cats}
\title{Example Cats}
\usage{
example_cats()
}
\value{
a data.frame containing example data for cats
}
\description{
Creates cat data with alignments for use in examples
}
\examples{
example_cats()

}
