% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalignment.R
\name{ggalignment}
\alias{ggalignment}
\title{ggalignment}
\usage{
ggalignment(
  alignment,
  background = ggalignment::dotted_lines_bg,
  alignment_coords = ggalignment::alignment_vals
)
}
\arguments{
\item{alignment_coords}{}
}
\value{

}
\description{
Creates a D&D alignment chart
}
