% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
rhs with lhs as the first argument
}
\description{
See \verb{magrittr::[\\\%>\\\%][magrittr::pipe]} for details.
}
\keyword{internal}
