% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_cats.R
\docType{data}
\name{align_cats}
\alias{align_cats}
\title{Align Cats}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\usage{
align_cats
}
\value{
align_cats, a data.frame containing example data
}
\description{
Creates align cats data for use in example code
}
\examples{
align_cats
}
\keyword{datasets}
