% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalignment.R
\name{ggalignment}
\alias{ggalignment}
\title{ggalignment}
\usage{
ggalignment(
  alignment,
  linetype = "dashed",
  linecolor = "black",
  font_family = "Arial",
  font_color = "black",
  background_color = "white",
  background_border = NA,
  background = ggalignment::dotted_lines_bg,
  alignment_coords = ggalignment::alignment_vals
)
}
\arguments{
\item{linetype}{the linetype for the box borders, which follows the ggplot2
allowable values for linetype for geom_rect() (e.g. blank, solid,
dashed, dotted, dotdash, longdash, twodash)}

\item{linecolor}{the color for the bounding boxes of the alignments, defaults
to black, and must be a named color such as "black"}

\item{background_color}{the background color for the entire plot, defaults to
white and must be a named color such as "white"}

\item{background_border}{the color of the solid-line bounding box on the entire
plot, defaults to NA and must be either NA or a named color such as "black"}
}
\value{

}
\description{
Creates a D&D alignment chart
}
