% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalignment.R
\name{ggalignment}
\alias{ggalignment}
\title{ggalignment}
\usage{
ggalignment(
  alignment,
  line_type = "dashed",
  line_color = "black",
  font_family = NULL,
  font_color = "black",
  background_color = "white",
  background_border = NA
)
}
\arguments{
\item{alignment}{a data.frame containing the data to be plotted, requiring
columns `img` (for image path) and `alignment`, and optionally `x` and `y`
specifying the coordinates for each image, where each box has coordinate limits
from -1 to 1 in both axes.}

\item{line_type}{the linetype for the box borders, which follows the ggplot2
allowable values for linetype for geom_rect() (e.g. blank, solid,
dashed, dotted, dotdash, longdash, twodash)}

\item{line_color}{the color for the bounding boxes of the alignments, defaults
to black, and must be a named color such as "black"}

\item{font_family}{the font family to be used on the alignment labels}

\item{font_color}{the font color to be used on the alignment labels}

\item{background_color}{the background color for the entire plot, defaults to
white and must be a named color such as "white"}

\item{background_border}{the color of the solid-line bounding box on the entire
plot, defaults to NA and must be either NA or a named color such as "black"}
}
\value{

}
\description{
Creates a D&D alignment chart
}
